Refactor expression-generation code into `ExprGenerator` (#90)
refactor expression-generation code into `ExprGenerator`

---------
